# Reduce the number of requests made by S3Files