--- a/s3v2/s3.go
+++ b/s3v2/s3.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"math"
 	"sync"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -42,9 +43,10 @@ type S3File struct {
 	putObjectInputOptions []func(*s3.PutObjectInput)
 
 	// read-related fields
-	readOpened bool
-	fileSize   int64
-	downloader *manager.Downloader
+	readOpened     bool
+	fileSize       int64
+	socket         io.ReadCloser
+	minRequestSize int
 
 	lock       sync.RWMutex
 	err        error
@@ -54,8 +56,9 @@ type S3File struct {
 }
 
 const (
-	rangeHeader       = "bytes=%d-%d"
-	rangeHeaderSuffix = "bytes=%d"
+	rangeHeader           = "bytes=%d-%d"
+	rangeHeaderSuffix     = "bytes=%d"
+	defaultMinRequestSize = math.MaxUint32
 )
 
 var (
@@ -106,35 +109,40 @@ func NewS3FileWriterWithClient(
 
 // NewS3FileReader creates an S3 FileReader, to be used with NewParquetReader
 func NewS3FileReader(ctx context.Context, bucket string, key string, cfgs ...*aws.Config) (source.ParquetFile, error) {
-	return NewS3FileReaderWithClientVersioned(ctx, s3.NewFromConfig(getConfig()), bucket, key, nil)
+	return NewS3FileReaderWithParams(ctx, S3FileReaderParams{
+		Bucket: bucket,
+		Key:    key,
+	})
 }
 
 // NewS3FileReaderVersioned creates an S3 FileReader for a versioned S3 object, to be used with NewParquetReader
 func NewS3FileReaderVersioned(ctx context.Context, bucket string, key string, version *string, cfgs ...*aws.Config) (source.ParquetFile, error) {
-	return NewS3FileReaderWithClientVersioned(ctx, s3.NewFromConfig(getConfig()), bucket, key, version)
+	return NewS3FileReaderWithParams(ctx, S3FileReaderParams{
+		Bucket:  bucket,
+		Key:     key,
+		Version: version,
+	})
 }
 
 // NewS3FileReaderWithClient is the same as NewS3FileReader but allows passing
 // your own S3 client
 func NewS3FileReaderWithClient(ctx context.Context, s3Client S3API, bucket string, key string) (source.ParquetFile, error) {
-	return NewS3FileReaderWithClientVersioned(ctx, s3Client, bucket, key, nil)
+	return NewS3FileReaderWithParams(ctx, S3FileReaderParams{
+		Bucket:   bucket,
+		Key:      key,
+		S3Client: s3Client,
+	})
 }
 
 // NewS3FileReaderWithClientVersioned is the same as NewS3FileReaderVersioned but allows passing
 // your own S3 client
 func NewS3FileReaderWithClientVersioned(ctx context.Context, s3Client S3API, bucket string, key string, version *string) (source.ParquetFile, error) {
-	s3Downloader := manager.NewDownloader(s3Client)
-
-	file := &S3File{
-		ctx:        ctx,
-		client:     s3Client,
-		downloader: s3Downloader,
-		BucketName: bucket,
-		Key:        key,
-		version:    version,
-	}
-
-	return file.Open(key)
+	return NewS3FileReaderWithParams(ctx, S3FileReaderParams{
+		Bucket:   bucket,
+		Key:      key,
+		S3Client: s3Client,
+		Version:  version,
+	})
 }
 
 // Seek tracks the offset for the next Read. Has no effect on Write.
@@ -163,6 +171,8 @@ func (s *S3File) Seek(offset int64, whence int) (int64, error) {
 
 	s.offset = offset
 	s.whence = whence
+
+	s.closeSocket()
 	return s.offset, nil
 }
 
@@ -172,7 +182,38 @@ func (s *S3File) Read(p []byte) (n int, err error) {
 		return 0, io.EOF
 	}
 
-	numBytes := len(p)
+	defer func() {
+		if err != nil {
+			s.closeSocket()
+		}
+	}()
+
+	if s.socket == nil {
+		err = s.openSocket(len(p))
+		if err != nil {
+			return 0, err
+		}
+	}
+	n, err = s.socket.Read(p)
+	// Because the chunk size is not infinite, we might hit the end of the socket while
+	// there's still data in the file. In this case, we close the socket so that the next
+	// read call will request a new one, and we return a nil error so that the caller
+	// will not think the file is done.
+	if err == io.EOF {
+		err = nil
+		s.closeSocket()
+	}
+	s.offset += int64(n)
+
+	return n, err
+}
+
+// openSocket issues a new GetObject request to retrieve the next chunk of data from the
+// object.
+func (s *S3File) openSocket(numBytes int) error {
+	if numBytes < s.minRequestSize {
+		numBytes = s.minRequestSize
+	}
 	getObjRange := s.getBytesRange(numBytes)
 	getObj := &s3.GetObjectInput{
 		Bucket:    aws.String(s.BucketName),
@@ -183,20 +224,19 @@ func (s *S3File) Read(p []byte) (n int, err error) {
 		getObj.Range = aws.String(getObjRange)
 	}
 
-	wab := manager.NewWriteAtBuffer(p)
-	bytesDownloaded, err := s.downloader.Download(s.ctx, wab, getObj)
+	out, err := s.client.GetObject(s.ctx, getObj)
 	if err != nil {
-		return 0, err
+		return err
 	}
+	s.socket = out.Body
+	return nil
+}
 
-	s.offset += bytesDownloaded
-	if buf := wab.Bytes(); len(buf) > numBytes {
-		// backing buffer reassigned, copy over some of the data
-		copy(p, buf)
-		bytesDownloaded = int64(len(p))
+func (s *S3File) closeSocket() {
+	if s.socket != nil {
+		s.socket.Close()
+		s.socket = nil
 	}
-
-	return int(bytesDownloaded), err
 }
 
 // Write len(p) bytes from p to the S3 data stream
@@ -234,6 +274,8 @@ func (s *S3File) Write(p []byte) (n int, err error) {
 func (s *S3File) Close() error {
 	var err error
 
+	s.closeSocket()
+
 	if s.pipeWriter != nil {
 		if err = s.pipeWriter.Close(); err != nil {
 			return err
@@ -266,15 +308,15 @@ func (s *S3File) Open(name string) (source.ParquetFile, error) {
 
 	// create a new instance
 	pf := &S3File{
-		ctx:        s.ctx,
-		client:     s.client,
-		downloader: s.downloader,
-		BucketName: s.BucketName,
-		Key:        name,
-		version:    s.version,
-		readOpened: s.readOpened,
-		fileSize:   s.fileSize,
-		offset:     0,
+		ctx:            s.ctx,
+		client:         s.client,
+		BucketName:     s.BucketName,
+		Key:            name,
+		version:        s.version,
+		readOpened:     s.readOpened,
+		fileSize:       s.fileSize,
+		minRequestSize: s.minRequestSize,
+		offset:         0,
 	}
 	return pf, nil
 }
@@ -367,7 +409,7 @@ func (s *S3File) getBytesRange(numBytes int) string {
 		end       int64
 	)
 
-	// Processing for unknown file size relies on the requestor to
+	// Processing for unknown file size relies on the requester to
 	// know which ranges are valid. May occur if caller is missing HEAD permissions.
 	if s.fileSize < 1 {
 		switch s.whence {
@@ -399,4 +441,52 @@ func (s *S3File) getBytesRange(numBytes int) string {
 
 	byteRange = fmt.Sprintf(rangeHeader, begin, end)
 	return byteRange
+}
+
+// S3FileReaderParams contains fields used to initialize and configure an S3File object
+// for reading.
+type S3FileReaderParams struct {
+	Bucket string
+	Key    string
+
+	// S3Client will be used to issue requests to S3. If not set, a new one will be
+	// created. Optional.
+	S3Client S3API
+	// Version is the version of the S3 object that will be read. If not set, the newest
+	// version will be read. Optional.
+	Version *string
+	// MinRequestSize controls the amount of data per request that the S3File will ask for
+	// from S3. Optional.
+	// A large MinRequestSize should improve performance and reduce AWS costs due to
+	// number of request. However, in some cases it may increase AWS costs due to data
+	// processing or data transfer. For best results, set it at or above the largest of
+	// the footer size and the biggest chunk size in the parquet file.
+	// S3File will not buffer a large amount of data in memory at one time, regardless
+	// of the value of MinRequestSize.
+	MinRequestSize int
+}
+
+// NewS3FileReaderWithParams creates an S3 FileReader for an object identified by and
+// configured using the S3FileReaderParams object.
+func NewS3FileReaderWithParams(ctx context.Context, params S3FileReaderParams) (source.ParquetFile, error) {
+	s3Client := params.S3Client
+	if s3Client == nil {
+		s3Client = s3.NewFromConfig(getConfig())
+	}
+
+	minRequestSize := params.MinRequestSize
+	if minRequestSize == 0 {
+		minRequestSize = defaultMinRequestSize
+	}
+
+	file := &S3File{
+		ctx:            ctx,
+		client:         s3Client,
+		BucketName:     params.Bucket,
+		Key:            params.Key,
+		version:        params.Version,
+		minRequestSize: minRequestSize,
+	}
+
+	return file.Open(params.Key)
 }